Me quede en hacer get de semanas con idUser y nombreSemana

+++ src/main/java/com/example/tfg/repositorios/SemanaUserRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.tfg.modelos.SemanaUser;
 
-public interface SemanaUserRepository extends JpaRepository<SemanaUser, Integer>{
+public interface SemanaUserRepository extends JpaRepository<SemanaUser, Long>{
 
 }
